Change DB types from integer to bigint

+++ api/migrations/20181216164536-create-constant.js
@@ -5,7 +5,7 @@ module.exports = {
       allowNull: false,
       autoIncrement: true,
       primaryKey: true,
-      type: Sequelize.INTEGER,
+      type: Sequelize.BIGINT,
     },
     key: {
       type: Sequelize.STRING,
